Rename gem accorgin rubygems naming convetion http://guides.rubygems.org/name-your-gem/ Fix #3
