+++ .rubocop.yml
@@ -14,3 +14,6 @@ AllCops:
 Metrics/LineLength:
   Max: 120
 
+Style/FileName:
+  Exclude:
+    - lib/activeadmin-reform.rb
