+++ lib/active_admin/reform.rb
@@ -1,0 +1,6 @@
+module ActiveAdmin
+  #
+  module Reform
+    require_relative 'reform/engine'
+  end
+end
