+++ lib/activeadmin-reform.rb
@@ -1,0 +1,1 @@
+require_relative 'active_admin/reform'
